feat: Add support for plotting motifs in Jupyter notebooks
This commit adds support for plotting motifs in Jupyter notebooks by checking if the code is running in a notebook environment. If it is, the notebook backend 'nbAgg' is used, otherwise the interactive backend 'TkAgg' is used. This allows for interactive plotting in notebooks and non-interactive plotting in other environments.